+++ spec/ratistics/collection_spec.rb
@@ -4,6 +4,116 @@ module Ratistics
 
   describe Collection do
 
+    context '#ascending?' do
+
+      it 'returns false for a nil sample' do
+        Collection.ascending?(nil).should be_false
+      end
+
+      it 'returns true for an empty sample' do
+        Collection.ascending?([].freeze).should be_true
+      end
+
+      it 'returns true for a one-element sample' do
+        Collection.ascending?([100].freeze).should be_true
+      end
+
+      it 'returns true for an ascending collection' do
+        Collection.ascending?([1, 2, 3, 4].freeze).should be_true
+      end
+
+      it 'returns false for a non-ascending collection' do
+        Collection.ascending?([1, 3, 2, 4].freeze).should be_false
+      end
+
+      it 'returns the correct value when given a block' do
+        sample = [
+          {:count => 11}, 
+          {:count => 12},
+          {:count => 13},
+          {:count => 14}
+        ].freeze
+
+        Collection.ascending?(sample){|item| item[:count]}.should be_true
+      end
+
+      context 'with ActiveRecord' do
+
+        before(:all) { Racer.connect }
+
+        specify do
+          sample = Racer.where('place > 0').order('place asc')
+          Collection.ascending?(sample){|r| r.place}.should be_true
+        end
+
+      end
+
+      context 'Hamster' do
+        
+        let(:list) { Hamster.list(3, 5, 6, 7, 8, 11, 15, 21, 22, 28, 30, 32, 33, 34, 40).freeze }
+        let(:vector) { Hamster.vector(3, 5, 6, 7, 8, 11, 15, 21, 22, 28, 30, 32, 33, 34, 40).freeze }
+
+        specify { Collection.ascending?(list).should be_true }
+
+        specify { Collection.ascending?(vector).should be_true }
+      end
+    end
+
+    context '#descending?' do
+
+      it 'returns false for a nil sample' do
+        Collection.descending?(nil).should be_false
+      end
+
+      it 'returns true for an empty sample' do
+        Collection.descending?([].freeze).should be_true
+      end
+
+      it 'returns true for a one-element sample' do
+        Collection.descending?([100].freeze).should be_true
+      end
+
+      it 'returns true for an descending collection' do
+        Collection.descending?([4, 3, 2, 1].freeze).should be_true
+      end
+
+      it 'returns false for a non-descending collection' do
+        Collection.descending?([1, 3, 2, 4].freeze).should be_false
+      end
+
+      it 'returns the correct value when given a block' do
+        sample = [
+          {:count => 21},
+          {:count => 20},
+          {:count => 19},
+          {:count => 18}
+        ].freeze
+
+        Collection.descending?(sample){|item| item[:count]}.should be_true
+      end
+
+      context 'with ActiveRecord' do
+
+        before(:all) { Racer.connect }
+
+        specify do
+          sample = Racer.where('place > 0').order('place desc')
+          Collection.descending?(sample){|r| r.place}.should be_true
+        end
+
+      end
+
+      context 'Hamster' do
+        
+        let(:list) { Hamster.list(40, 34, 33, 32, 30, 28, 22, 21, 15, 11, 8, 7, 6, 5, 3).freeze }
+        let(:vector) { Hamster.vector(40, 34, 33, 32, 30, 28, 22, 21, 15, 11, 8, 7, 6, 5, 3).freeze }
+
+        specify { Collection.descending?(list).should be_true }
+
+        specify { Collection.descending?(vector).should be_true }
+      end
+    end
+
     context '#binary_search' do
 
       let(:sample) do